issue-01: all functionality is done

+++ web/src/main/java/controller/ItemController.java
@@ -1,21 +1,21 @@
 package controller;
 
-import dto.ItemDto;
+import exception.ErrorMessage;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import model.Availability;
 import model.Category;
 import model.Item;
 import model.Manufacturer;
+import model.Subcategory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
-import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.servlet.ModelAndView;
 import repository.CategoryRepository;
 import repository.ItemRepository;
@@ -51,23 +51,28 @@ public class ItemController {
     @PostMapping("/itemSelection")
     public String processSelectedItem(@RequestParam(required = false) String param,
             @ModelAttribute Item selectedItem) {
-        if (param.equals("DELETE")) {
-            itemRepository.delete(selectedItem.getId());
-            return "redirect:/successPage";
-        } else if (param.equals("UPDATE")) {
-            return "redirect:/updateItem/" + selectedItem.getId();
+        Item item = itemRepository.findOne(selectedItem.getId());
+        if (item != null) {
+            if (param.equals("DELETE")) {
+                itemRepository.delete(selectedItem.getId());
+                return "redirect:/success";
+            } else if (param.equals("UPDATE")) {
+                return "redirect:/updateItem/" + selectedItem.getId();
+            }
+            return "redirect:/createItem";
         }
-        return "redirect:/createItem";
+        return "redirect:/error/" + ErrorMessage.NO_ITEM.getMessage();
     }
 
-    @Transactional
     @GetMapping("/createItem")
     public ModelAndView createItem() {
         ModelAndView modelAndView = new ModelAndView("createItem");
         List<Category> categories = categoryRepository.findAll();
+        ArrayList<Subcategory> subcategories = new ArrayList<>();
+        categories.forEach(c-> subcategories.addAll(c.getSubcategories()));
         List<Manufacturer> manufacturers = manufacturerRepository.findAll();
         Availability availability = Availability.ABSENT;
-        modelAndView.addObject("categories", categories);
+        modelAndView.addObject("subcategories", subcategories);
         modelAndView.addObject("manufacturers", manufacturers);
         modelAndView.addObject("availability", availability);
         modelAndView.addObject("newItem", new Item());
@@ -77,7 +82,7 @@ public class ItemController {
     @PostMapping("/createItem")
     public String createItem(@ModelAttribute(name = "newItem") Item newItem) {
         itemRepository.save(newItem);
-        return "redirect:/successPage";
+        return "redirect:/success";
     }
 
     @GetMapping("/updateItem/{id}")
@@ -87,10 +92,10 @@ public class ItemController {
         Item newItem = new Item();
         newItem.setId(id);
         List<Manufacturer> manufacturers = manufacturerRepository.findAll();
-        List<Category> categories = categoryRepository.findAll();
+        List<Subcategory> subcategories = oldItem.getSubcategory().getCategory().getSubcategories();
         modelAndView.addObject("oldItem", oldItem);
         modelAndView.addObject("manufacturers", manufacturers);
-        modelAndView.addObject("categories", categories);
+        modelAndView.addObject("subcategories", subcategories);
         modelAndView.addObject("availabilities", Arrays.asList(Availability.values()));
         modelAndView.addObject("newItem", newItem);
         return modelAndView;
@@ -99,6 +104,6 @@ public class ItemController {
     @PostMapping("/updateItem/{id}")
     public String showItemUpdated(@ModelAttribute Item newItem) {
         itemRepository.save(newItem);
-        return "redirect:/successPage";
+        return "redirect:/success";
     }
 }
